[streaming] Renamed test cases for window and grouped window invokable to be consistent with the invokable names

+++ flink-addons/flink-streaming/flink-streaming-core/src/main/java/org/apache/flink/streaming/api/windowing/extractor/FieldsFromTuple.java
@@ -46,7 +46,7 @@ public class FieldsFromTuple implements Extractor<Tuple, double[]> {
 	public double[] extract(Tuple in) {
 		double[] out = new double[indexes.length];
 		for (int i = 0; i < indexes.length; i++) {
-			out[i] = in.getField(indexes[i]);
+			out[i] = (Double) in.getField(indexes[i]);
 		}
 		return out;
 	}
